add version; module name is more readable; move files into lib/subdirectory;  format (whitespace)
