+++ lib/aasm-statechart/version.rb
@@ -1,0 +1,3 @@
+module AASM_StateChart
+  VERSION = '1.0.2'
+end
